Create min method size threshold for RF
Prevent trigger a refactor diligence warning for increases in method
size when method is still very small.

Run with MIN_METHOD_SIZE=6 for the moment.
